refactor: rename kebab-case
