--- src/lib/github.ts
@@ -1,4 +1,0 @@
-// GitHub releases data is fetched at build time by the githubReleasesPlugin in vite.config.ts
-// and exposed as the virtual module "virtual:releases".
-// This file is kept as documentation of the data source.
-export {};
