+++ src/routes/about.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { motion } from "motion/react";
-import { SEO } from "@/components/SEO";
+import { SEO } from "@/scripts/seo";
 
 export const Route = createFileRoute("/about")({
   component: AboutPage,
